Refactoring notify and checknewmessages into the shared consumer module
