+++ edgerouter/go.mod
@@ -5,7 +5,7 @@ go 1.22.2
 require (
 	github.com/google/uuid v1.6.0
 	github.com/libp2p/zeroconf/v2 v2.2.0
-	github.com/maritimeconnectivity/MMS/consumers v0.0.0
+	github.com/maritimeconnectivity/MMS/consumer v0.0.0-00010101000000-000000000000
 	github.com/maritimeconnectivity/MMS/mmtp v0.0.0
 	github.com/maritimeconnectivity/MMS/utils v0.0.0
 	nhooyr.io/websocket v1.8.11
@@ -24,4 +24,4 @@ replace github.com/maritimeconnectivity/MMS/mmtp => ../mmtp
 
 replace github.com/maritimeconnectivity/MMS/utils => ../utils
 
-replace github.com/maritimeconnectivity/MMS/consumers => ../consumers
+replace github.com/maritimeconnectivity/MMS/consumer => ../consumer
